refactor: Atomic 디렉토리구조로 변경
