+++ src/components/atom/main/sidebar/Instagram.tsx
@@ -1,0 +1,6 @@
+import { SocialIcon } from "./SocialIconStyles";
+import instagram from "../../../../assets/images/instagram_logo.svg";
+
+const Instagram = () => <SocialIcon src={instagram} alt="Instagram" />;
+
+export default Instagram;
