+++ src/components/atom/main/sidebar/Logo.tsx
@@ -1,0 +1,6 @@
+import { Logo } from "./LogoStyles";
+import logo from "../../../../assets/images/homepage_logo_top.svg";
+
+const SidebarLogo = () => <Logo src={logo} alt="Logo" />;
+
+export default SidebarLogo;
