+++ src/components/atom/main/sidebar/LogoStyles.ts
@@ -1,0 +1,9 @@
+import tw from "twin.macro";
+import styled from "styled-components";
+
+export const Logo = styled.img`
+  ${tw`
+    w-28
+    m-2
+  `}
+`;
