+++ src/components/atom/main/sidebar/SocialIconStyles.ts
@@ -1,0 +1,8 @@
+import tw from "twin.macro";
+import styled from "styled-components";
+
+export const SocialIcon = styled.img`
+  ${tw`
+    w-7 h-7
+  `}
+`;
