+++ src/components/atom/main/sidebar/Youtube.tsx
@@ -1,0 +1,6 @@
+import { SocialIcon } from "./SocialIconStyles";
+import youtube from "../../../../assets/images/youtube_logo.svg";
+
+const Youtube = () => <SocialIcon src={youtube} alt="Youtube" />;
+
+export default Youtube;
