+++ src/components/molecules/main/sidebar/SocialIcons.tsx
@@ -1,0 +1,12 @@
+import { SocialIcons } from "./SocialIconsStyles";
+import Instagram from "../../../atom/main/sidebar/Instagram";
+import Youtube from "../../../atom/main/sidebar/Youtube";
+
+const Social = () => (
+  <SocialIcons>
+    <Instagram />
+    <Youtube />
+  </SocialIcons>
+);
+
+export default Social;
