+++ src/components/molecules/main/sidebar/SocialIconsStyles.ts
@@ -1,0 +1,12 @@
+import tw from "twin.macro";
+import styled from "styled-components";
+
+export const SocialIcons = styled.div`
+  ${tw`
+    flex 
+    flex-col
+    justify-between
+    space-y-10
+    mb-20
+  `}
+`;
