+++ src/components/organism/main/Main.tsx
@@ -1,0 +1,5 @@
+import { MainPageContainer } from "./MainStyles";
+
+const Main = () => <MainPageContainer></MainPageContainer>;
+
+export default Main;
