+++ src/components/organism/main/MainStyles.ts
@@ -1,0 +1,12 @@
+import tw from "twin.macro";
+import styled from "styled-components";
+
+export const MainPageContainer = styled.div`
+  ${tw`
+    flex
+    flex-col
+    bg-cover
+    bg-no-repeat
+    w-screen
+  `}
+`;
