+++ src/pages/HomePageStyles.ts
@@ -1,0 +1,9 @@
+import tw from "twin.macro";
+import styled from "styled-components";
+
+export const HomePageContainer = styled.div`
+  ${tw`
+    flex 
+    h-screen
+  `}
+`;
